add poisson calc and almost complete

+++ app/src/main/java/com/example/statisticcalculationapp/Result.kt
@@ -1,9 +1,11 @@
 package com.example.statisticcalculationapp
 
+import java.math.BigInteger
+
 // TODO: convert data class to normal optional class
 data class Result(
     var summary: String,
-    val attemptsNumber: Int,
-    val successNumber: Int,
+    val attemptsNumber: BigInteger,
+    val successNumber: BigInteger,
     val probabilityOfSuccess: Double,
 )
